Renamed some files/added stub functions for use cases
The renaming I did was just to make clear what the boundary and controller were. I'm quite tired writing this so sorry if this is obvious but I think the entities are on the AWS and database end of thing so we won't have those classes. Organization is a bit messy, I can fix that another time but just wanted to make sure this was set up somewhat. This project doesn't have the professor's React hack used to pass a model back and forth and that relevant information if that ends up being needed (although we don't have a canvas so I don't know?). After logging in we need a way to keep track of and hold the credentials on the client side so maybe some entity might be useful for that (so naturally the authentication page needs to be completed). Sorry for the late night ramble!